:recycle: refactor: 경로명 오타 수정
